move js into src
