+++ src/careen.js
@@ -1,5 +1,3 @@
-'use strict';
-
 const path = require('path');
 const config = require('./config');
