+++ src/config.js
@@ -1,5 +1,3 @@
-'use strict';
-
 require('dotenv').config();
 
 module.exports = {
